Create README.md
developed a unity 3D game. wherein a car in constant motion must avoid multiple obstacles in the path, 
car moves right and left using ADWS keys. changing the color of the obstacles when the car hits obstacles.
created a variable called "score" and initialized it to 100, each time the car hits the obstacle reduces the score by 10.
when the score value is zero, stops the car in motion.